Renamed files, separated parts in to modules to allow for greater overlap with background development, fixed batchsize and compile mode setter bug.
