+++ cosipy/response/NFWorkerState.py
@@ -1,0 +1,3 @@
+worker_device = None
+density_module = None
+area_module = None
